+++ tests/test_singlepoint_cli.py
@@ -6,7 +6,7 @@ from ase.io import read
 from typer.testing import CliRunner
 import yaml
 
-from janus_core.cli import app
+from janus_core.cli.cli import app
 from tests.utils import assert_log_contains, read_atoms
 
 DATA_PATH = Path(__file__).parent / "data"
